+++ app/src/main/java/edu/nextstep/camp/calculator/domain/Operate.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.calculator
+package edu.nextstep.camp.calculator.domain
 
 interface Operate {
