+++ app/src/main/java/edu/nextstep/camp/calculator/domain/Operator.kt
@@ -1,4 +1,4 @@
-package edu.nextstep.camp.calculator
+package edu.nextstep.camp.calculator.domain
 
 data class Operator(val value: String) {
